fixes axios version / ensures errors are accessible
change handle response error function to api file
