+++ client/package.json
@@ -46,7 +46,7 @@
     "@tanstack/react-query": "^4.2.1",
     "@tanstack/react-table": "8.5.1",
     "autoprefixer": "10.2.5",
-    "axios": "^1.3.4",
+    "axios": "1.3.4",
     "chroma-js": "2.1.2",
     "classnames": "2.3.1",
     "d3-array": "3.0.2",
