+++ client/src/pages/data/scenarios/[scenarioId]/interventions/[interventionId]/edit.tsx
@@ -9,7 +9,7 @@ import CleanLayout from 'layouts/clean';
 import InterventionForm from 'containers/interventions/form';
 import BackLink from 'components/back-link/component';
 import Loading from 'components/loading';
-import { handleResponseError } from 'utils/handle-response-error';
+import { handleResponseError } from 'services/api';
 
 import type { InterventionFormData } from 'containers/interventions/types';
